feat: communication wearable / handheld
Using a common module for data and drawables handheld and
wearable now talks via WearApi data item and messages.
Received weather data is stored in shared preferences on
the wearable.
With missing weather data the watchfaces shows a message
and a button to refresh that triggers a sync on the
handheld.